Moved Entities in new directory structure
