+++ src/Model/Entities/Agents/Agent.java
@@ -1,0 +1,18 @@
+package Model.Entities.Agents;
+
+import java.awt.Color;
+
+import Model.Entities.Entity;
+
+/**
+ * This is a abstract class defining general behavior and attributes of all Agents.
+ */
+public class Agent extends Entity
+{
+
+    public Agent(int id, Color color)
+    {
+        super(id, color);
+    }
+    
+}
